+++ xmake/core/project/config.lua
@@ -237,6 +237,51 @@ function config.clear()
     config._CONFIGS = nil
 end
 
+-- the current mode is belong to the given modes?
+function config.is_mode(...)
+
+    -- get the current mode
+    local mode = config.get("mode")
+    if not mode then return false end
+
+    -- exists this mode?
+    for _, m in ipairs(table.join(...)) do
+        if m and type(m) == "string" and m == mode then
+            return true
+        end
+    end
+end
+
+-- the current platform is belong to the given platforms?
+function config.is_plat(...)
+
+    -- get the current platform
+    local plat = config.get("plat")
+    if not plat then return false end
+
+    -- exists this platform? and escape '-'
+    for _, p in ipairs(table.join(...)) do
+        if p and type(p) == "string" and plat:find("^" .. p:gsub("%-", "%%-") .. "$") then
+            return true
+        end
+    end
+end
+
+-- the current platform is belong to the given architectures?
+function config.is_arch(...)
+
+    -- get the current architecture
+    local arch = config.get("arch")
+    if not arch then return false end
+
+    -- exists this architecture? and escape '-'
+    for _, a in ipairs(table.join(...)) do
+        if a and type(a) == "string" and arch:find("^" .. a:gsub("%-", "%%-") .. "$") then
+            return true
+        end
+    end
+end
+
 -- dump the configure
 function config.dump()