+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [#158](https://github.com/tboox/xmake/issues/158): Support CUDA Toolkit and Compiler
 * Add `set_tools` and `add_tools` apis to change the toolchains for special target
 * Add builtin rules: `mode:debug` and `mode:release`
+* Add `is_mode`, `is_arch` and `is_plat` builtin apis in the custom scripts
 
 ### Changes
 
@@ -450,6 +451,7 @@
 * [#158](https://github.com/tboox/xmake/issues/158): 增加对Cuda编译环境的支持
 * 添加`set_tools`和`add_tools`接口为指定target目标设置编译工具链
 * 添加内建规则：`mode:debug`和`mode:release`
+* 添加`is_mode`, `is_arch` 和`is_plat`内置接口到自定义脚本域
 
 ### 改进